+++ src/frontend/app/shared-sdoc/resolver/sdoc-album.resolver.ts
@@ -5,7 +5,7 @@ import {IdValidationRule} from '../../../shared/search-commons/model/forms/gener
 import {ResolvedData, ResolverError} from '../../../shared/angular-commons/resolver/resolver.utils';
 import {LogUtils} from '../../../shared/commons/utils/log.utils';
 import {SDocSearchForm, SDocSearchFormFactory} from '../../../shared/sdoc-commons/model/forms/sdoc-searchform';
-import {IdCsvValidationRule} from '../../../../shared/search-commons/model/forms/generic-validator.util';
+import {IdCsvValidationRule} from '../../../shared/search-commons/model/forms/generic-validator.util';
 import {SDocAlbumService} from '../services/sdoc-album.service';
 
 @Injectable()