+++ src/frontend/environments/environment.beta.ts
@@ -5,6 +5,8 @@ export const environment = {
     backendApiBaseUrl: 'http://www.michas-ausflugstipps.de/mytbbeta/api/v1/',
     tracksBaseUrl: 'http://www.michas-ausflugstipps.de/mytbbeta/api/assets/trackstore/',
     picsBaseUrl: 'http://www.michas-ausflugstipps.de/mytbbeta/api/static/picturestore/',
+    defaultSearchTypes: 'route,location,track,trip,news',
+    emptyDefaultSearchTypes: 'route,location,track,trip,news,image',
     useAssetStoreUrls: true,
     sdocWritable: false,
     sdocActionTagWritable: false,